Refactor: 페이지 네이밍을 명시적으로 변경(Setting -> SplitTypeChange)
