+++ src/App.tsx
@@ -6,9 +6,9 @@ import {
   HOME_PAGE,
   NOT_FOUND_PAGE,
   SURVEY_PAGE,
-  SETTING_PAGE,
+  SPLIT_TYPE_CHAGNE_PAGE,
 } from '@/consts/route';
-import { Home, Survey, ExerciseChoice, ExerciseRoutine, NotFound, Setting } from '@/pages';
+import { Home, Survey, ExerciseChoice, ExerciseRoutine, NotFound, SplitTypeChange } from '@/pages';
 import { Loading } from '@/components';
 
 const App = () => {
@@ -20,7 +20,7 @@ const App = () => {
           <Route path={SURVEY_PAGE} component={Survey} />
           <Route path={EXERCISE_CHOICE_PAGE} component={ExerciseChoice} />
           <Route path={EXERCISE_ROUTINE_PAGE} component={ExerciseRoutine} />
-          <Route path={SETTING_PAGE} component={Setting} />
+          <Route path={SPLIT_TYPE_CHAGNE_PAGE} component={SplitTypeChange} />
           <Route path={NOT_FOUND_PAGE} component={NotFound} />
           <Redirect to={NOT_FOUND_PAGE} />
         </Switch>
