+++ src/consts/route.ts
@@ -3,4 +3,4 @@ export const SURVEY_PAGE = '/survey';
 export const EXERCISE_ROUTINE_PAGE = '/exercise-routine';
 export const EXERCISE_CHOICE_PAGE = '/exercise-choice';
 export const NOT_FOUND_PAGE = '/not-found';
-export const SETTING_PAGE = '/setting';
+export const SPLIT_TYPE_CHAGNE_PAGE = '/setting/split-type';
