+++ src/pages/SplitTypeChange/splitTypeChange.module.scss
