+++ src/pages/index.ts
@@ -2,5 +2,5 @@ export { default as Home } from './Home/Home';
 export { default as Survey } from './Survey/Survey';
 export { default as ExerciseChoice } from './ExerciseChoice/ExerciseChoice';
 export { default as ExerciseRoutine } from './ExerciseRoutine/ExerciseRoutine';
-export { default as Setting } from './Setting/Setting';
+export { default as SplitTypeChange } from './SplitTypeChange/SplitTypeChange';
 export { default as NotFound } from './NotFound/NotFound';
